refactor(pm): split resolver driver into focused submodules
Decompose the 1003-line `demand/driver.rs` into a `driver/` module and
collapse the per-edge resolution decision tree.

- Extract `plan_edge`/`EdgeStep`: the four near-identical resolution arms
  (semver / version-cache / full-manifest-cache / versions-list) fold into
  one pure decision over the store. `run_main_loop_bfs` drops 411 -> 239
  lines and the failure/skip/cache logic becomes testable in isolation.
- Split by concern: `plan.rs` (decisions, no I/O), `schedule.rs` (enqueue
  demand/prefetch jobs), `pipeline.rs` (execute fetches + apply results),
  `run.rs` (BFS loop + graph mutation). `mod.rs` is declarations only.

No behaviour change: 174 + 10 ruborist tests pass, clippy clean (default
and all-features).
